% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2lambda.R
\name{df2lambda}
\alias{df2lambda}
\title{Selection probabilities of the regularization model using the elastic-net penalty with multiresponse gaussian family (MNET).}
\usage{
df2lambda(
  x,
  y,
  alpha,
  family,
  type.multinomial,
  seq.df,
  penalty.factor = NULL,
  verbose = TRUE
)
}
\description{
The penalized regression model using elastic-net penalty with multiresponse gaussian family (MNET) aims to solve the following problem:
}
