% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet.R
\name{mnet}
\alias{mnet}
\title{Selection probabilities of the regularization model using the elastic-net penalty with multiresponse gaussian family (MNET).}
\usage{
mnet(
  x,
  y,
  seq.alpha = NULL,
  seq.lambda = NULL,
  K = 100,
  psub = 0.5,
  standardize.response = TRUE,
  penalty.factor = NULL,
  ...
)
}
\arguments{
\item{x}{Either R object or file path can be considered. A genotype data is not a data.frame but a matrix with dimension \code{p} by \code{(n+11)}. It is formatted by hapmap which has (rs, allele, chr, pos) in the first four(1-4) columns, (strand, assembly, center, protLSID, assayLSID, panel, Qcode) in the following seven(5-11) columns. If NULL, user can choose a path in interactive use.}

\item{y}{Either R object or file path can be considered. A phenotype data is an \code{n} by \code{p} matrix. Since the first some columns can display attributes of the phenotypes, you should enter the arguments, y.col and y.id.col, which represent the columns of phenotypes to be analyzed and the column of sample ID. If NULL, user can choose a path in interactive use.}

\item{seq.alpha}{Default is "object". If \code{input.type} is "object", obejects of genotype/phenotype will be entered, and if "path", paths of genotype/phenotype will be enterd. If you want to use an object, you have to make sure that the class of each column of genotype data is equal to "character".}

\item{seq.lambda}{The columns of phenotypes. At most 4 phenotypes can be considered, because the plot of them will be fine. Default is 2.}

\item{K}{The column of sample ID in the phenotype data file. Default is 1.}

\item{psub}{If TRUE. phenotypes are converted to be normal-shape using box-cox transformation when all phenotypes are positive.}

\item{standardize.response}{If TRUE. phenotypes are converted to be normal-shape using box-cox transformation when all phenotypes are positive.}

\item{penalty.factor}{If TRUE. phenotypes are converted to be normal-shape using box-cox transformation when all phenotypes are positive.}
}
\value{
A data.frame with each column representing the selection probability for each phenotype
}
\description{
The penalized regression model using elastic-net penalty with multiresponse gaussian family (MNET) aims to solve the following problem:
\deqn{ \frac{1}{2} \| Y - X \beta \|_F^2 +  \lambda [ \frac{1-\alpha}{2} ||\beta||_F^2 + \alpha \sum_{j=1}^p ||\beta_j||_2 ]}
}
\references{
Zou, H., & Hastie, T. (2005). Regularization and variable selection via the elastic net. Journal of the royal statistical society: series B (statistical methodology), 67(2), 301-320.
Meinshausen, N., & Bühlmann, P. (2010). Stability selection. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 72(4), 417-473.
Simon, N., Friedman, J., & Hastie, T. (2013). A blockwise descent algorithm for group-penalized multiresponse and multinomial regression. arXiv preprint arXiv:1311.6529.
Kim, K., Koo, J., & Sun, H. (2020). An empirical threshold of selection probability for analysis of high-dimensional correlated data. Journal of Statistical Computation and Simulation, 1-12.
}
\author{
Kipoong Kim \href{mailto:kkp7700@gmail.com}{kkp7700@gmail.com}
}
