% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet-package.R
\docType{package}
\name{unet-package}
\alias{unet-package}
\title{A package for calculating the selection probabilities of a provided regularization approach in the situation with multivariate phenotypes.}
\value{
A data.frame of selection probabilities for each phenotypes.
}
\description{
unet
}
\details{
In reality, there are many types of phenotypes such as continuous, ordinal, binary, category, etc.
For such phenotypes, there have been many researches to solve a multivariate regression problem.
For an analysis of high-dimensional genomic data with a single phenotype, the penalized regression has been widely used in many fields because it can fit a regression model for the multiple covariates at a time while accomodating correlations between predictors.
However, most of them focused on continuous phenotypes only, and the regularization approaches can depend on the tuning parameter selection which can be a challenging problem.
This package provides the selection probability for the case of multivariate phenotypes, which can prioritize variables given in a model regardless of tuning parameter selection.
Additionally, the empirical threshold is also provided to control the false positives(errors).
Thresholds vary with the expected number of false positives to be specified by the user.
Currently, this package contains three methods of unified elasitc-net (UNET), unified elastic-net with df-aggregating (UNET-df) and their comparing method, multiresponse elastic-net (MNET).
}
\references{
Zou, H., & Hastie, T. (2005). Regularization and variable selection via the elastic net. Journal of the royal statistical society: series B (statistical methodology), 67(2), 301-320.
Meinshausen, N., & Bühlmann, P. (2010). Stability selection. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 72(4), 417-473.
Simon, N., Friedman, J., & Hastie, T. (2013). A blockwise descent algorithm for group-penalized multiresponse and multinomial regression. arXiv preprint arXiv:1311.6529.
Kim, K., Koo, J., & Sun, H. (2020). An empirical threshold of selection probability for analysis of high-dimensional correlated data. Journal of Statistical Computation and Simulation, 1-12.
}
\keyword{"selection}
\keyword{gwas}
\keyword{multivariate}
\keyword{probabilities"}
\keyword{regularization}
