% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{unet}
\alias{unet}
\title{U-Net: Convolutional Networks for Biomedical Image Segmentation}
\usage{
unet(input_shape, num_classes = 1, dropout = 0.5, filters = 64,
  num_layers = 4, output_activation = "sigmoid")
}
\arguments{
\item{input_shape}{Dimensionality of the input (integer) not including the
samples axis. Must be lenght 3 numeric vector.}

\item{num_classes}{Number of classes.}

\item{dropout}{Dropout rate applied.}

\item{filters}{Number of filters of the first convolution.}

\item{num_layers}{Number of layers in the encoder.}

\item{output_activation}{Activation in the output layer.}
}
\description{
U-Net: Convolutional Networks for Biomedical Image Segmentation
}
